06_Projects/config_data_files/autogen_ollama_requirements.txt
pyautogen>=0.2.0
aiohttp>=3.9.0
ccxt>=4.0.0
python-dotenv>=1.0.0
